requirements.txt
Pillow
pandas
loguru
yolox
